chore: Merge back into `syslog.rs`
No changes beyond relocating the code into a single file.
